+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 Changelog
 =========
 
+# 2.1.0
+
+- nackAll requeue argument is true by default, as per documentation
+- acking/nacking/rejecting an already acked message kills the channel with an unknown delivery tag error (406)
+
 # 2.0.0
 
 - drop node 12 support
 - convert to esm with exports for node
-- smqp@8
+- bump `smqp@8`
 - lint some
 
 # 1.0.0